integrations-examples/livekit-agent/transcriber/pyproject.toml
[project]
name = "livekit-example-transcriber"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "livekit-agents[gladia]>=1.2.6",
]
